rename Fixes #113
